@@ -1,7 +1,7 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId "org.jun1or.dwebview_android"
         minSdkVersion 17
@@ -19,17 +19,15 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
+    implementation "androidx.appcompat:appcompat:1.1.0"
+    implementation "androidx.core:core:1.2.0"
 //    implementation project(':dwebview')
-    implementation 'org.jun1or.baselib:dwebview:1.0.0'
-    implementation 'org.jun1or.baselib:util:1.0.0'
+    implementation 'org.jun1or.baselib:dwebview:1.0.1'
+    implementation 'org.jun1or.baselib:util:1.0.1'
     implementation 'io.reactivex.rxjava2:rxandroid:2.0.1'
     implementation 'io.reactivex.rxjava2:rxjava:2.1.16'
-    implementation 'com.android.support:design:28.0.0'
-    implementation('com.github.bumptech.glide:glide:4.7.1') {
-        exclude group: 'com.android.support'
-    }
-    annotationProcessor 'com.github.bumptech.glide:compiler:4.7.1'
-    implementation 'org.jun1or.baselib:imgsel:1.0.0'
+    implementation('com.github.bumptech.glide:glide:4.11.0')
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.11.0'
+    implementation 'org.jun1or.baselib:imgsel:1.0.1'
 
 }
