@@ -2,12 +2,13 @@ package org.jun1or.dwebview_android;
 
 import android.graphics.Color;
 import android.os.Bundle;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+
 
 import org.jun1or.dwebview.fragment.WebFragment;
-import com.istrong.dwebview_android.R;
 
 
 public class JavascriptCallNativeActivity extends AppCompatActivity {
