@@ -2,13 +2,14 @@ package org.jun1or.dwebview_android;
 
 import android.graphics.Color;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
 
 import org.jun1or.dwebview.fragment.WebFragment;
-import com.istrong.dwebview_android.R;
+
 
 public class WebFragmentPreloadActivity extends AppCompatActivity {
 
