@@ -1,10 +1,12 @@
 package org.jun1or.dwebview_android;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
 import android.view.View;
 
+import androidx.annotation.Nullable;
+
 import org.jun1or.dwebview.fragment.WebFragment;
+
 
 public class JWebFragment extends WebFragment {
 
