@@ -1,9 +1,15 @@
 package org.jun1or.dwebview.callback;
 
 /**
- * Created by du on 16/12/31.
+ * @author cwj
+ * @date 16/12/31
  */
 
 public interface OnReturnValue<T> {
+    /**
+     * 返回值
+     *
+     * @param retValue
+     */
     void onValue(T retValue);
 }
